+++ src/main/java/hs/os/skaterverwaltung/entity/ISkaterRepository.java
@@ -3,11 +3,11 @@ import java.util.Collection;
 import hs.os.skaterverwaltung.control.dto.SkaterDTO;
 
 public interface ISkaterRepository {
-    public Skater get(int id);
+    public Skater get(long id);
     public Collection<Skater> getAll();
     public Skater add (SkaterDTO dto);
-    public Skater edit (int id, SkaterDTO dto);
-    public boolean delete (int id);
+    public Skater edit (long id, SkaterDTO dto);
+    public boolean delete (long id);
 
 
 }
